Simplify the provisioner fix per review
Drop the engine-side recording of provisioner/connection references as
resource dependencies: hooks are opaque named callbacks in every Pulumi
SDK and imply no dependency edges, and the graph walk already orders
the referent's registration (and settles its outputs) before the
dependent resource binds its hooks.

Replace the evaluator struct with the original free eval helpers, which
now deep-unmark their results; whether output must be suppressed is
decided by a single configSensitive check over the provisioner's
configuration, which also matches suppression being a property of the
whole config block rather than of the attributes a runner happens to
read.
