+++ .changes/unreleased/runtime-bug-fixes-333.yaml
@@ -1,5 +1,5 @@
 kind: bug-fixes
-body: Fix a panic when a `provisioner` or `connection` block references another resource's outputs or a sensitive value, and record those references as resource dependencies
+body: Fix a panic when a `provisioner` or `connection` block references another resource's outputs or a sensitive value
 time: 2026-07-10T10:03:38Z
 custom:
     Component: runtime
