+++ pkg/provisioner/runtime/local_exec.go
@@ -23,6 +23,9 @@ import (
 	"runtime"
 
 	"github.com/hashicorp/hcl/v2"
+	"github.com/zclconf/go-cty/cty"
+
+	"github.com/pulumi-labs/pulumi-hcl/pkg/hcl/eval"
 )
 
 // when and on_failure are stripped by the parser before this body reaches us.
@@ -42,9 +45,7 @@ func runLocalExec(ctx context.Context, spec *Spec, hclCtx *hcl.EvalContext) erro
 		return fmt.Errorf("local-exec: %s", diags.Error())
 	}
 
-	ev := &evaluator{hclCtx: hclCtx}
-
-	command, err := ev.evalString(content, "command")
+	command, err := evalString(content, "command", hclCtx)
 	if err != nil {
 		return err
 	}
@@ -52,12 +53,12 @@ func runLocalExec(ctx context.Context, spec *Spec, hclCtx *hcl.EvalContext) erro
 		return fmt.Errorf("local-exec: command must be non-empty")
 	}
 
-	workingDir, err := ev.evalString(content, "working_dir")
+	workingDir, err := evalOptionalString(content, "working_dir", hclCtx)
 	if err != nil {
 		return err
 	}
 
-	interpreter, err := ev.evalStringSlice(content, "interpreter")
+	interpreter, err := evalStringSlice(content, "interpreter", hclCtx)
 	if err != nil {
 		return err
 	}
@@ -65,15 +66,17 @@ func runLocalExec(ctx context.Context, spec *Spec, hclCtx *hcl.EvalContext) erro
 		interpreter = defaultInterpreter()
 	}
 
-	environment, err := ev.evalStringMap(content, "environment")
+	environment, err := evalStringMap(content, "environment", hclCtx)
 	if err != nil {
 		return err
 	}
 
-	quiet, err := ev.evalBool(content, "quiet")
+	quiet, err := evalOptionalBool(content, "quiet", hclCtx)
 	if err != nil {
 		return err
 	}
+
+	sensitive := configSensitive(content, hclCtx)
 
 	// TF appends command as the final positional arg of interpreter,
 	// e.g. ["/bin/sh", "-c", "echo hi"].
@@ -87,10 +90,10 @@ func runLocalExec(ctx context.Context, spec *Spec, hclCtx *hcl.EvalContext) erro
 			cmd.Env = append(cmd.Env, k+"="+v)
 		}
 	}
-	if ev.sensitive {
+	if sensitive {
 		fmt.Fprintln(os.Stderr, suppressedOutputMsg)
 	}
-	if quiet || ev.sensitive {
+	if quiet || sensitive {
 		cmd.Stdout = io.Discard
 		cmd.Stderr = io.Discard
 	} else {
@@ -109,4 +112,124 @@ func defaultInterpreter() []string {
 		return []string{"cmd", "/C"}
 	}
 	return []string{"/bin/sh", "-c"}
+}
+
+// configSensitive reports whether any attribute of the provisioner's
+// configuration evaluates to a value carrying the sensitive mark, in which
+// case the provisioner's output must be suppressed so the value cannot leak
+// through logging.
+func configSensitive(content *hcl.BodyContent, hclCtx *hcl.EvalContext) bool {
+	for _, attr := range content.Attributes {
+		val, diags := attr.Expr.Value(hclCtx)
+		if diags.HasErrors() {
+			continue
+		}
+		_, marks := val.UnmarkDeep()
+		if _, ok := marks[eval.SensitiveMark]; ok {
+			return true
+		}
+	}
+	return false
+}
+
+// The eval helpers deep-unmark every result: the engine tracks cross-resource
+// dependencies and sensitivity as cty marks, and marked values panic in
+// AsString and friends.
+
+func evalString(content *hcl.BodyContent, name string, hclCtx *hcl.EvalContext) (string, error) {
+	attr, ok := content.Attributes[name]
+	if !ok {
+		return "", nil
+	}
+	val, diags := attr.Expr.Value(hclCtx)
+	if diags.HasErrors() {
+		return "", fmt.Errorf("evaluating %s: %s", name, diags.Error())
+	}
+	val, _ = val.UnmarkDeep()
+	if !val.IsKnown() || val.IsNull() {
+		return "", nil
+	}
+	if val.Type() != cty.String {
+		return "", fmt.Errorf("%s must be a string, got %s", name, val.Type().FriendlyName())
+	}
+	return val.AsString(), nil
+}
+
+func evalOptionalString(content *hcl.BodyContent, name string, hclCtx *hcl.EvalContext) (string, error) {
+	return evalString(content, name, hclCtx)
+}
+
+func evalOptionalBool(content *hcl.BodyContent, name string, hclCtx *hcl.EvalContext) (bool, error) {
+	attr, ok := content.Attributes[name]
+	if !ok {
+		return false, nil
+	}
+	val, diags := attr.Expr.Value(hclCtx)
+	if diags.HasErrors() {
+		return false, fmt.Errorf("evaluating %s: %s", name, diags.Error())
+	}
+	val, _ = val.UnmarkDeep()
+	if !val.IsKnown() || val.IsNull() {
+		return false, nil
+	}
+	if val.Type() != cty.Bool {
+		return false, fmt.Errorf("%s must be a bool, got %s", name, val.Type().FriendlyName())
+	}
+	return val.True(), nil
+}
+
+func evalStringSlice(content *hcl.BodyContent, name string, hclCtx *hcl.EvalContext) ([]string, error) {
+	attr, ok := content.Attributes[name]
+	if !ok {
+		return nil, nil
+	}
+	val, diags := attr.Expr.Value(hclCtx)
+	if diags.HasErrors() {
+		return nil, fmt.Errorf("evaluating %s: %s", name, diags.Error())
+	}
+	val, _ = val.UnmarkDeep()
+	if !val.IsKnown() || val.IsNull() {
+		return nil, nil
+	}
+	if !val.CanIterateElements() {
+		return nil, fmt.Errorf("%s must be a list of strings", name)
+	}
+	out := make([]string, 0, val.LengthInt())
+	it := val.ElementIterator()
+	for it.Next() {
+		_, v := it.Element()
+		if v.Type() != cty.String {
+			return nil, fmt.Errorf("%s elements must be strings", name)
+		}
+		out = append(out, v.AsString())
+	}
+	return out, nil
+}
+
+func evalStringMap(content *hcl.BodyContent, name string, hclCtx *hcl.EvalContext) (map[string]string, error) {
+	attr, ok := content.Attributes[name]
+	if !ok {
+		return nil, nil
+	}
+	val, diags := attr.Expr.Value(hclCtx)
+	if diags.HasErrors() {
+		return nil, fmt.Errorf("evaluating %s: %s", name, diags.Error())
+	}
+	val, _ = val.UnmarkDeep()
+	if !val.IsKnown() || val.IsNull() {
+		return nil, nil
+	}
+	if !val.CanIterateElements() {
+		return nil, fmt.Errorf("%s must be a map of strings", name)
+	}
+	out := make(map[string]string, val.LengthInt())
+	it := val.ElementIterator()
+	for it.Next() {
+		k, v := it.Element()
+		if k.Type() != cty.String || v.Type() != cty.String {
+			return nil, fmt.Errorf("%s keys and values must be strings", name)
+		}
+		out[k.AsString()] = v.AsString()
+	}
+	return out, nil
 }
